requirements.txt
pandas==1.3.5
matplotlib==3.8.2
psycopg2-binary==2.9.9
pyspark==3.5.4
confluent-kafka==2.8.0